優化首頁統計邏輯，加入啟動腳本

+++ DouBanAPI/template-main/src/main/java/com/step/template/main/vo/MovieItem.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class MovieItem {
     private Double rating; // Assuming rating is a double
     private Integer id; // Assuming id is a long
     private  String goodRateStr;
-    private  Integer goodCount;
+    private Long goodCount;
     private  Integer year;
 }
